upgrade-test Move classes of com.liferay.portal.upgrade.v7_0_3 package to module
