+++ modules/apps/portal/portal-upgrade-test/src/testIntegration/java/com/liferay/portal/upgrade/v7_0_3/test/UpgradeSQLServerTest.java
@@ -12,24 +12,28 @@
  * details.
  */
 
-package com.liferay.portal.upgrade.v7_0_3;
+package com.liferay.portal.upgrade.v7_0_3.test;
 
+import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.portal.kernel.dao.db.DB;
 import com.liferay.portal.kernel.dao.db.DBManagerUtil;
 import com.liferay.portal.kernel.dao.db.DBType;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
-import com.liferay.portal.upgrade.BaseUpgradeDBColumnSizeTestCase;
+import com.liferay.portal.upgrade.test.BaseUpgradeDBColumnSizeTestCase;
+import com.liferay.portal.upgrade.v7_0_3.UpgradeSQLServer;
 
 import org.junit.Assume;
 import org.junit.BeforeClass;
 import org.junit.ClassRule;
 import org.junit.Rule;
+import org.junit.runner.RunWith;
 
 /**
  * @author Preston Crary
  */
-public class UpgradeSybaseTest extends BaseUpgradeDBColumnSizeTestCase {
+@RunWith(Arquillian.class)
+public class UpgradeSQLServerTest extends BaseUpgradeDBColumnSizeTestCase {
 
 	@ClassRule
 	@Rule
@@ -40,28 +44,28 @@ public class UpgradeSybaseTest extends BaseUpgradeDBColumnSizeTestCase {
 	public static void setUpClass() {
 		DB db = DBManagerUtil.getDB();
 
-		Assume.assumeTrue(DBType.SYBASE.equals(db.getDBType()));
+		Assume.assumeTrue(DBType.SQLSERVER.equals(db.getDBType()));
 	}
 
 	@Override
 	protected String getCreateTestTableSQL() {
-		return "create table TestTable (testTableId decimal(20, 0) not null " +
-			"primary key, testValue varchar(1000) null)";
+		return "create table TestTable (testTableId int not null primary " +
+			"key, testValue varchar(2000) null)";
 	}
 
 	@Override
 	protected int getInitialSize() {
-		return 1000;
+		return 2000;
 	}
 
 	@Override
 	protected String getTypeName() {
-		return "varchar";
+		return "nvarchar";
 	}
 
 	@Override
-	protected UpgradeSybase getUpgradeProcess() {
-		return new UpgradeSybase();
+	protected UpgradeSQLServer getUpgradeProcess() {
+		return new UpgradeSQLServer();
 	}
 
 }
